% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundCatGauss.R
\name{boundCatGauss}
\alias{boundCatGauss}
\title{Lower bound for multiple categorical variables}
\usage{
boundCatGauss(X, model, prior)
}
\arguments{
\item{X}{NxD matrix data}

\item{model}{Model parameters}

\item{prior}{Model parameters}
}
\value{
Value of lower bound
}
\description{
Lower bound for multiple categorical variables
}
