% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound.R
\name{boundGauss}
\alias{boundGauss}
\title{Lower bound for multivariate Gaussian}
\usage{
boundGauss(X, model, prior, indep)
}
\arguments{
\item{X}{NxD matrix data}

\item{model}{Model parameters}

\item{prior}{Model parameters}

\item{indep}{Independent covariates (boolean)}
}
\value{
Value of lower bound
}
\description{
Lower bound for multivariate Gaussian
}
