% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundIndGauss.R
\name{boundIndGauss}
\alias{boundIndGauss}
\title{Lower bound for independent Gaussian}
\usage{
boundIndGauss(X, model, prior, verbose)
}
\arguments{
\item{X}{NxD matrix data}

\item{model}{Model parameters}

\item{prior}{Model parameters}
}
\value{
Value of lower bound
}
\description{
Lower bound for independent Gaussian
}
