% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundMulGauss.R
\name{boundMulGauss}
\alias{boundMulGauss}
\title{Lower bound for multivariate Gaussian}
\usage{
boundMulGauss(X, model, prior)
}
\arguments{
\item{X}{NxD matrix data}

\item{model}{Model parameters}

\item{prior}{Model parameters}
}
\value{
Value of lower bound
}
\description{
Lower bound for multivariate Gaussian
}
