% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundSelGauss.R
\name{boundSelGauss}
\alias{boundSelGauss}
\title{Lower bound for independent Gaussian}
\usage{
boundSelGauss(X, model, prior)
}
\arguments{
\item{X}{NxD matrix data}

\item{model}{Model parameters}

\item{prior}{Model parameters}
}
\value{
Value of lower bound
}
\description{
Lower bound for independent Gaussian
}
