% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundUniGauss.R
\name{boundUniGauss}
\alias{boundUniGauss}
\title{Lower bound for univariate Gaussian}
\usage{
boundUniGauss(X, model, prior)
}
\arguments{
\item{X}{NxD matrix data}

\item{model}{Model parameters}

\item{prior}{Model parameters}
}
\value{
Value of lower bound
}
\description{
Lower bound for univariate Gaussian
}
