% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conve.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence of variational algorithm}
\usage{
check_convergence(L, iter, tol, maxiter, verbose)
}
\arguments{
\item{L}{Lower bound}

\item{iter}{Iteration number}

\item{tol}{Tolerance}

\item{maxiter}{Maximum number of iterations}

\item{verbose}{Boolean flag that, if TRUE, prints the outcome of the convergence check}
}
\value{
Boolean flag that indicates if algorithm has converged
}
\description{
Check convergence of variational algorithm
}
