% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expec.R
\name{expeCat}
\alias{expeCat}
\title{Expectation of log(pi)}
\usage{
expeCat(alpha, k)
}
\arguments{
\item{alpha}{Dirichlet hyperparameter.}

\item{k}{Mixture component of interest.}
}
\description{
Expectation of log(pi)
}
