% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectCat.R
\name{expectCatGauss}
\alias{expectCatGauss}
\title{Expectation step for mixture of categorical variables}
\usage{
expectCatGauss(X, model)
}
\arguments{
\item{X}{NxD data matrix.}

\item{model}{Model parameters.}
}
\value{
Updated model parameters.
}
\description{
Expectation step for mixture of categorical variables
}
