% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectIndGauss.R
\name{expectIndGauss}
\alias{expectIndGauss}
\title{Expectation step for mixture of independent Gaussians}
\usage{
expectIndGauss(X, model)
}
\arguments{
\item{X}{NxD data matrix.}

\item{model}{Model parameters.}
}
\value{
Updated model parameters.
}
\description{
Expectation step for mixture of independent Gaussians
}
