% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectMulGauss.R
\name{expectMulGauss}
\alias{expectMulGauss}
\title{Expectation step for mixture of multivariate Gaussians}
\usage{
expectMulGauss(X, model)
}
\arguments{
\item{X}{NxD data matrix.}

\item{model}{Model parameters.}
}
\value{
Updated model parameters.
}
\description{
Expectation step for mixture of multivariate Gaussians
}
