% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectSelGauss.R
\name{expectSelGauss}
\alias{expectSelGauss}
\title{Expectation step for mixture of independent Gaussians}
\usage{
expectSelGauss(X, model, null)
}
\arguments{
\item{X}{NxD data matrix.}

\item{model}{Model parameters.}
}
\value{
Updated model parameters.
}
\description{
Expectation step for mixture of independent Gaussians
}
