% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expec.R
\name{expeGauss}
\alias{expeGauss}
\title{Update log(Rho) for independent Gaussians}
\usage{
expeGauss(X, D, N, v_k, m_k, M_k, beta_k, logW_k)
}
\arguments{
\item{X}{NxD data matrix.}

\item{D}{Number of covariates.}

\item{N}{Number of observations.}

\item{v_k}{Precision hyperparameter.}

\item{m_k}{Mean hyperparameter.}

\item{M_k}{Precision hyperparameter.}

\item{beta_k}{Mean hyperparameter.}

\item{logW_k}{log(1/det(M)), not needed if indep is TRUE).}
}
\description{
Update log(Rho) for independent Gaussians
}
