% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize responisibility matrix}
\usage{
init(X, K, init = F, verbose = T)
}
\arguments{
\item{X}{NxD data matrix}

\item{K}{Number of mixture components}

\item{init}{Initialisation method (optional). If it is a vector, it is interpreted as the vector of initial cluster allocations.
If it is a string, it is interpreted as the name of the clustering algorithm used for the initialisation (only "kmeans")
available at the moment).}

\item{verbose}{Boolean flag which, if TRUE, prints the type of initialization}
}
\value{
Resp NxK matrix of responsibilities
}
\description{
This function allows you to do variational Bayesian inference for Gaussian mixture.
}
