% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expec.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Use the log sum exp trick to improve numerical stability}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Vector of values}
}
\description{
Use the log sum exp trick to improve numerical stability
}
