% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expec.R
\name{logRho_to_Resp}
\alias{logRho_to_Resp}
\title{Transform log(Rho) matrix to responsibilities matrix}
\usage{
logRho_to_Resp(logRho)
}
\arguments{
\item{logRho}{log(Rho).}
}
\value{
Matrix of responsibilities.
}
\description{
Transform log(Rho) matrix to responsibilities matrix
}
