% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logUniB.R
\name{logUniB}
\alias{logUniB}
\title{Compute univariate equivalent of logB}
\usage{
logUniB(W, nu)
}
\arguments{
\item{W}{Da rivedere}

\item{nu}{Da rivedere}
}
\description{
Compute univariate equivalent of logB
}
