% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxim.R
\name{maximizeIndGauss}
\alias{maximizeIndGauss}
\title{Maximization step for independent Gaussian}
\usage{
maximizeIndGauss(X, model, prior)
}
\arguments{
\item{X}{NxD data matrix}

\item{model}{Model parameters}

\item{prior}{Prior parameters}
}
\value{
Updated model parameters
}
\description{
Maximization step for independent Gaussian
}
