% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeMulGauss.R
\name{maximizeMulGauss}
\alias{maximizeMulGauss}
\title{Maximization step for multivariate Gaussian}
\usage{
maximizeMulGauss(X, model, prior)
}
\arguments{
\item{X}{NxD data matrix}

\item{model}{Model parameters}

\item{prior}{Prior parameters}
}
\value{
Updated model parameters
}
\description{
Maximization step for multivariate Gaussian
}
