% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeSalGauss.R
\name{maximizeSalGauss}
\alias{maximizeSalGauss}
\title{Maximization step for categorical variables}
\usage{
maximizeSalGauss(X, model, prior)
}
\arguments{
\item{X}{NxD data matrix}

\item{model}{Model parameters}

\item{prior}{Prior parameters}
}
\value{
Updated model parameters
}
\description{
Maximization step for categorical variables
}
