% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeUniGauss.R
\name{maximizeUniGauss}
\alias{maximizeUniGauss}
\title{Maximization step for univariate Gaussian}
\usage{
maximizeUniGauss(X, model, prior)
}
\arguments{
\item{X}{NxD data matrix}

\item{model}{Model parameters}

\item{prior}{Prior parameters}
}
\value{
Updated model parameters
}
\description{
Maximization step for univariate Gaussian
}
