% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessCat.R
\name{preprocessCategoricalDataset}
\alias{preprocessCategoricalDataset}
\title{Preprocessing matrix containing categorical variables}
\usage{
preprocessCategoricalDataset(X)
}
\arguments{
\item{X}{NxD matrix}
}
\description{
Preprocessing matrix containing categorical variables
}
\references{
Inspiration came from code contained in the R package 'mixdir'
}
