% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxim.R
\name{update_invW}
\alias{update_invW}
\title{Update M = inv(W)}
\usage{
update_invW(X, Resp_k, M0_k, beta0_k, m0_k, xbar_k)
}
\arguments{
\item{X}{NxD data matrix.}

\item{Resp_k}{Vector of responsibilities for mixture component k.}

\item{M0_k}{Prior precision for mixture component k.}

\item{beta0_k}{Prior beta for mixture component k.}

\item{m0_k}{Prior m for mixture component k.}

\item{xbar_k}{Weighted mean for mixture component k.}
}
\description{
Update M = inv(W)
}
