% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxim.R
\name{update_m}
\alias{update_m}
\title{Update mean vector}
\usage{
update_m(beta0, m0, Nk, xbar, beta)
}
\arguments{
\item{beta0}{Prior beta.}

\item{m0}{Prior m.}

\item{Nk}{Number of observations in component k.}

\item{xbar}{Weighted mean of component k.}

\item{beta}{Parameter for precision of mu.}
}
\description{
Update mean vector
}
