% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxim.R
\name{update_xbar}
\alias{update_xbar}
\title{Update xbar}
\usage{
update_xbar(X, Resp)
}
\arguments{
\item{X}{NxD data matrix.}

\item{Resp}{Vector of responsibilities for the mixture component of interest.}
}
\description{
Update xbar
}
