% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectSalGauss.R
\name{updateSalGauss}
\alias{updateSalGauss}
\title{Expectation step for mixture of independent Gaussians}
\usage{
updateSalGauss(X, model, prior)
}
\arguments{
\item{X}{NxD data matrix.}

\item{model}{Model parameters.}

\item{prior}{Prior hyperparameters.}
}
\value{
Updated model parameters.
}
\description{
Expectation step for mixture of independent Gaussians
}
