% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimix.R
\name{vimix}
\alias{vimix}
\title{Variational Bayesian inference for unsupervised clustering}
\usage{
vimix(X, K, prior, indep = F, init = "kmeans", select = F,
  tol = 1e-04, maxiter = 2000, verbose = F)
}
\arguments{
\item{X}{NxD data matrix.}

\item{K}{(Maximum) number of clusters.}

\item{prior}{Prior parameters (optional).}

\item{indep}{Booleand indicator. If TRUE, the features are considered to be independent. Default is FALSE.}

\item{init}{Initialisation method (optional). If it is a vector, it is interpreted as the vector of initial
cluster allocations. If it is a string, it is interpreted as the name of the clustering algorithm used for
the initialisation (only "kmeans" and "random") available at the moment).}

\item{select}{Boolean flag. If TRUE, variable selection is used. Default is FALSE.}

\item{tol}{Tolerance on lower bound. Default is 10e-20.}

\item{maxiter}{Maximum number of iterations of the VB algorithm. Default is 2000.}

\item{verbose}{Boolean flag which, if TRUE, prints the iteration numbers. Default is FALSE.}
}
\value{
A list containing L, the lower bound at each step of the algorithm, label, a vector containing the
cluster labels, model, a list containing the trained model structure.
}
\description{
Variational Bayesian inference for unsupervised clustering
}
\examples{
library(mvtnorm)
data <- rbind(rmvnorm(100,c(-3,0)), rmvnorm(100,c(3,0)))
output <- vimix(data, 2)
}
\references{
Bishop, C.M., 2006. Pattern recognition and machine learning. Springer.
}
\author{
Alessandra Cabassi \email{alessandra.cabassi@mrc-bsu.cam.ac.uk}
}
