% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimix.R
\name{vimix_multiGauss}
\alias{vimix_multiGauss}
\title{Variational Bayesian inference for unsupervised clustering, mixture of multivariate Gaussians}
\usage{
vimix_multiGauss(X, K, prior, init = "kmeans", tol = 1e-19,
  maxiter = 2000, verbose = F)
}
\arguments{
\item{X}{NxD data matrix.}

\item{K}{(Maximum) number of clusters.}

\item{prior}{Prior parameters (optional).}

\item{init}{Initialisation method (optional). If it is a vector, it is interpreted as the vector of initial
cluster allocations. If it is a string, it is interpreted as the name of the clustering algorithm used for
the initialisation (only "kmeans" and "random") available at the moment).}

\item{tol}{Tolerance on lower bound. Default is 10e-20.}

\item{maxiter}{Maximum number of iterations of the VB algorithm. Default is 2000.}

\item{verbose}{Boolean flag which, if TRUE, prints the iteration numbers. Default is FALSE.}
}
\value{
A list containing L, the lower bound at each step of the algorithm, label, a vector containing the
cluster labels, model, a list containing the trained model structure.
}
\description{
Variational Bayesian inference for unsupervised clustering, mixture of multivariate Gaussians
}
\examples{
## Load a dataset containing 200 2-dimensional data points
data <- as.matrix(read.csv(system.file("extdata", "example1-data.csv", package = "vimix"),
row.names = 1))

## Use variational inference for mixture of Gaussians to find clusters
output <- vimix_multiGauss(data, 2)
}
\references{
Bishop, C.M., 2006. Pattern recognition and machine learning. Springer.
}
\author{
Alessandra Cabassi \email{alessandra.cabassi@mrc-bsu.cam.ac.uk}
}
