% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimix.R
\name{vimixCatGauss}
\alias{vimixCatGauss}
\title{Variational Bayesian inference for unsupervised clustering, mixture of categorical variables}
\usage{
vimixCatGauss(X, K, prior, init = "kmeans", tol = 1e-19,
  maxiter = 2000, verbose = F)
}
\arguments{
\item{X}{NxD data matrix.}

\item{K}{(Maximum) number of clusters.}

\item{prior}{Prior parameters (optional).}

\item{tol}{Tolerance on lower bound. Default is 10e-20.}

\item{maxiter}{Maximum number of iterations of the VB algorithm. Default is 2000.}

\item{verbose}{Boolean flag which, if TRUE, prints the iteration numbers. Default is FALSE.}
}
\value{
A list containing L, the lower bound at each step of the algorithm, label, a vector containing the
cluster labels, model, a list containing the trained model structure.
}
\description{
Variational Bayesian inference for unsupervised clustering, mixture of categorical variables
}
\references{
Bishop, C.M., 2006. Pattern recognition and machine learning. Springer.
}
\author{
Alessandra Cabassi \email{alessandra.cabassi@mrc-bsu.cam.ac.uk}
}
