% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expec.R
\name{expectUniGauss}
\alias{expectUniGauss}
\title{Expectation step for mixture of multivariate Gaussians}
\usage{
expectUniGauss(X, model)
}
\arguments{
\item{X}{NxD data matrix.}

\item{model}{Model parameters.}
}
\value{
Updated model parameters.
}
\description{
Expectation step for mixture of multivariate Gaussians
}
